PLC-TestProject: Add precompiled arm64 binaries
Soon we want to test on arm64, too. For this we the same precompiled
PLC-TestProjects as we already have for the amd64 service container.

NOTE: We don't use the TwinCAT PLC platform names "TwinCAT OS (ARMV8-A)".
Instead we use "arm64" suffix, which we already use together with
BHF_CI_ARCH, already.
Another problem with the original TwinCAT names are spaces, Docker COPY
is allergic to them.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>