ci: build docker image for service container
We want to replace the outdated "build-linuxum:patrickbr-docker-adslib"
image used for the service container as ads-server. Our TC/LUR test
pipeline already creates containers, but we can't use them directly. Our
AdsLibTest and AdsLibTestRef unittests rely on a static server NetId
192.168.0.231.1.1, so we have to customize the upstream container a bit.
We use that opportunity to copy in our PLC-TestProject, too.

The only thing missing is a valid license. We have to add it dynamically
when we use the new container in a follow up commit.

NOTE: We put the docker build into a new ".pre" stage. The gitlab-ci.yml
is already messy as is, but if we would add the necessary "needs:
parallel: matrix:" rules to the test jobs we would drive at least one
developer insance!

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>