ci: use our own container as ads-server
Previously, we used a hacked together container with an ancient Linux
usermode runtime as "ads-server" service container. Now, we have much
better options. We can use a slightly modified version of the latest
TC/LUR test container. This will make it easy to add arm64, too.

The new container provides much more telemetry, which we should make
availabe in GitLab artifacts.

NOTE: This change influences the systemids on kvm-build02 and
kvm-build03. Additionally, we will drop the few_vcpus tag and run on
even more machines. For this we add the systemids for kvm-build01-08.
Since this will become pretty unreadable, we add a new first column to
our table. This doesn't effect our grep matches.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>