ci: start testing on arm64, too
Finally, we are able to run TcAdsTool on arm64 and test against an arm64
service container.
For this we have to add the systemids of kvm-build04/05.
The gitlab-ci.yml becomes a bit messy, but until GitLab allows variables
in needs[1] I see no better approach...
We might merge the build and test jobs again? Would reduce the size of
the yaml significantly.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>