ci: install test dependencies into build container, too
With this we can use a single container image for both the build and
test jobs. This is the first step to merge them.

One of the biggest problems with our .gitlab-ci.yml is to feed the build
artifacts from the parallel:matrix jobs into the corresponding test
jobs. As our build is pretty simple and we don't even need the
intermediate build artifacts for archival. It makes much more sense to
merge build and test jobs into a single GitLab jobs, which will result
in a _much_ simpler YAML file.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>