ci: merge TC/BSD build and test job
It is easy to run the build on TC/BSD, too. So lets merge build and test
too and get rid of the weird artifact dependency.

NOTE: We don't archive any artefacts, well we would have to copy them
from the VM, which is possible but tedious. At the moment it doesn't
seem necessary, so we just ignore them.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>