clang-format: reformat the entire codebase
We are tired of uncrustify. In every version there are major changes,
which require large changes. Just now we want to upgrade to Debian
trixie(13) and would have to reformat large parts of our codebase.

Well, we don't like the old codestyle anyways. So lets try a new tool
and style. The Linux kernel uses clang-format and has a history of
stability. So lets try that.

Additionally, we use the opportunity to get normalized lineendings. We
ran dos2unix on all the code files.

NOTE: This is the last step of our conversion.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>