tools/run-uncrustify.sh: switch to clang-format
We are tired of uncrustify. In every version there are major changes,
which require large changes. Just now we want to upgrade to Debian
trixie(13) and would have to reformat large parts of our codebase.

Well, we don't like the old codestyle anyways. So lets try a new tool
and style. The Linux kernel uses clang-format and has a history of
stability. So lets try that. Oh and while we are at it we fix the find.
Well, we actually never formatted the header files before...

NOTE: This breaks the codestyle job in CI. We will fix it in the next
commit, but want to keep this change separated.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>