debian: update version to 113.0.29-1
We intend to release adstool packages for Debian Trixie. For our other
tools, we are changing the version number to 13.*.* to enhance
visibility. However, since we have already used git tags ranging from v1
to- v28.1 for this tool, seeing a version like v13.0.29 might be
somewhat confusing. This is why we are adding an additional 100 to the
version number. You might wonder why we chose to add 100 instead of 20.
Well, just because!

adstool (113.0.29-1) trixie; urgency=medium
.
  * ci: build with Debian trixie(13)
  * standalone/AmsRouter::AddRoute(): preserve exception type

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>