README: update instructions for clients with local TwinCAT router
For many users it is to hard to look into .gitlab-ci.yml and find out
how to build on different platforms with different flavours so we add
more documentation to the README.

On TC/BSD and TC/LUR we can use the new native meson files to build with
TwinCAT support. Additionally, we explain the difference between
"adstool" and "tcadstool".

Suggested-by: Birger Evenburg <B.Evenburg@beckhoff.com>
Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>