standalone/AdsLib: allow AdsWrite for zero length buffer
We have at least one usecase for AdsSyncWriteReqEx() called with a
nullptr buffer and zero length. ADSIGRP_EXTERNALTIME[1] uses this and
encodes the actual command within the IndexOffset.

AdsSyncReadWriteReqEx2() already uses the same pattern to check the
pointer only if the bufferLength is non zero. So we adapt the same
behaviour for AdsSyncWriteReqEx().

[1] https://infosys.beckhoff.com/content/1033/tc3_grundlagen/6313305611.html

Fixes: https://github.com/Beckhoff/ADS/issues/264
Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>