AmsResponse: fix deadlock in AmsResponse::Wait()
Previously, we used 0xFFFFFFFF as a special error code to detect if the
response processing was completed or not. Well, it turns out the system
service might respond with this code as well, so we would end up in a
deadlock. Even worse, that error code is read from the network, so we
could easily deadlock an AdsClient remotely.

To fix this a modern approach would be to just use std::optional[1].
However, we support old platforms and our minimal standard is C++14.

We can fix our problem with a simple bool, which doesn't increase any
compiler requirements. So we do the simple thing and get rid of our
special define.

[1] https://en.cppreference.com/w/cpp/utility/optional.html

Reported-by: Phil Kahlert <P.Kahlert@beckhoff.com>
Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>